Split informant in two
